requirements.txt
pyTelegramBotAPI
google-generativeai
scipy
pandas
requests
python-dotenv
aiohttp
